[Dsl] generalize packagingOptions support
The packagingOptions properties are sets of strings; in Groovy, these
properties can be assigned to using a literal list (in KotlinScript,
we require a (Mutable)Set).

The methods to update the property are singular and take one argument,
which they add to any existing set; there is no varargs method which
adds all of its arguments.

This means that a number of the existing (Groovy) test files are
written with invalid syntax; update those files to express correct Dsl
syntax.  Testing the contents of written-out files reveals that the
Groovy writer emits invalid Dsl; log a bug and test what we can,
adding expected files and commented-out test code for those tests
which would currently fail.  KotlinScript tests for this are currently
completely disabled, because all of them require parsing and
appropriate handling of MutableSet operators (at least setOf /
mutableSetOf).

Bug: 141842964
Bug: 141970574
Bug: 142114586
Test: existing tests updated
Change-Id: I9ecab64f1732970cd0255da9544b7dedbebff867