[Dsl] generalize lintOptions support
There are some boolean properties which are named differently in
Kotlin (isFoo) and Groovy (foo).  These are straightforwardly
handled.  Other boolean properties (e.g. htmlReport) are named the
same in Groovy and Kotlin.

There are some File properties (lintConfig, htmlOutput, textOutput,
xmlOutput) with a Groovy-only setter named the same.  Among those,
textOutput is special in that it also has a String setter, which
affects the writability of the Kotlin property.  There are
Single-Argument Method Transforms in play here, too, and the odd
textOutput behaviour in Kotlin exposes a use of the model to determine
whether to use assignment or a method call; we can't yet remove that
because of elements not yet updated to the new way of working, so note
a TODO.

Finally, there are some Set<String> properties with slightly odd
semantics.  Assignment to those properties is allowed, but adds the
values in the right-hand-side to the existing set.  (This is
currently unimplemented and untested in the Dsl parser).

Bug: 141842964
Bug: 141970574
Bug: 142114586
Test: existing tests updated, strengthened, run on .kts
Change-Id: I48d28edc18c42859f0f17bc1f8405ab82110a804