[Dsl,.kts] another check for assignment or methodcall
In KotlinScript, Single-Argument Method Transforms mean that we
sometimes end up in createDslMethodCall for properties.  We must at
that point check for model names and whether properties are writeable
or we should prefer method calls, because otherwise we might end up
emitting Dsl with invalid syntax: for example lintOptions.textOutput
is unusual in that it has a read-only property and a one-argument
setter in Kotlin, so the default assignment syntax is wrong.

(There is no analogous issue in Groovy, where a DslMethodCall is
always emitted as a Groovy command.)

Bug: 141970574
Test: in following CL
Change-Id: Iabbbbddde1686adcb456c2bf6dbfa707a2320a1c