Cache activities list from AndroidManifestIndex
Currently DefaultActivityLocator queries AndroidManifestIndex to
build a list of activities available to a module each time we execute
a run config that launches the default activity. We'd like to also
leverage the index for run config validation, which is called very
frequently on the UI thread.

As a first step in this direction, this CL caches the list of activities
in the AndroidFacet's user date so that we aren't constantly querying the
index when it isn't necessary to do so. We use the
MergedManifestModificationTracker for cache invalidation.

Bug: 143693655
Test: DefaultActivityLocatorTest
Change-Id: I24b4e6c71815ed782f719d226cabcfb900eac105