Push jar using transport file manager
This will allow AppInspectionPipelineConnection instances to push
pre-configured jars to host device with existing transport file utility.

The jars will need to be packaged into the release resource directory.
This means we need to modify how studio is built whenever we need to
add a new jar or update an existing jar.

On device, the jar will be located in the studio resource directory. The
same directory where perfa and other profiler libraries are located.

For example our current androidx library lives in:
prebuilts/tools/common/m2/repository/androidx/inspection/inspection/1.0.0-SNAPSHOT

When studio is built, it will be copied to plugins/android/resources.

When inspection connection is made, it will be copied to the profiler
resource directory.

An example of how a deployable file could be initialized:
DeployableFile.Builder("inspection-1.0.0-20191024.120903-1.jar")
  .setReleaseDir("plugins/android/resources/inspection")
  .setDevDir("../../prebuilts/tools/common/m2/repository/androidx/inspection/inspection/1.0.0-SNAPSHOT")
  .build()

Test: AppInspectionPipelineConnectionTest.launchInspector
Bug: 142526985
Change-Id: I275c08677f5bc5fd21519afdd5092fa473637b73