BLeak: respect DoNotTrace for static fields
Also, simplify ClassStaticsExpander by making it inherit directly from
Expander rather than DefaultObjectExpander. It only used this for
expanding the instance fields of Class objects (which it only did in
expandCorrespondingEdge, not in expand, so this was broken anyway). I
don't think there's much of interest there. This change makes it much
easier to reason about this expander's behavior.

Move FieldLabel from DefaultObjectExpander to Expander.

Change-Id: I4ec313b1cd4b6742180d0f639954e7c629bcb57e