[Dsl,.kts] test unknown values
At some point in the past, the special handling of one-argument method
calls in extra property initializers was removed or fixed.  The
remaining problem in unknown values handling was a failure to properly
track the Psi of dot-qualified method call initializers, which we can
now handle by explicitly passing in the Psi for method call argument
lists at call sites of getMethodCall().

Bug: 130224064
Test: GradlePropertyModel unknownValues* tests
Change-Id: I362c5196e4e8951553bb9e6c172878b766703d1b