[Dsl,.kts] support set properties
Because of the compile-time typed nature of KotlinScript, assignment
to properties must always be done with values of the right
compile-time type; this means that some of the convenience of
automatic coercion in Groovy is lost.

In particular, some properties are of type MutableSet, and while in
Groovy the user can assign a literal list to those properties, in
KotlinScript the object must be of type MutableSet.

Since most of the Dsl/Model will be oblivious to the set-like nature
of these objects, continue representing them as
GradleDslExpressionList objects, with a field to indicate whether they
should be treated as a set.  Populate this field either from the
model, if the property we are dealing with is known to be a set, or
from the parser, if we have parsed a set constructor such as
mutableSetOf(), and use this field when writing out.

Bug: 143193213
Test: new AndroidModel addAndApplyListElements for .kts
Test: existing tests updated
Change-Id: I548dc1436e9eee378813fefb9618e8f3084f9c8b