[Dsl] enforce consistency in the modelMap collector
Specifically the consistency that if we are describing a
property (null arity) our description should relate to property
semantics, and if we are describing a method of some kind (numeric
arity) the description should relate to method semantics.  Extend the
key function in the collector to perform these consistency checks
before returning the key.

This reveals two errors in the current Dsl descriptions.  Fix those.

Bug: none
Test: existing, additional enforced data consistency
Change-Id: Iae47719e8a976cf89abf37c4b8b280e7911b31cc