Do not set SCREEN_COMPOSE_ONLY as default SceneMode
When the Layout Editor creates a new NlDesignSurface it will not set the
scene mode by default so the last one will be loaded. This needs
revisiting but, for now, Compose is not a mode that is allowed in the
"normal" Layout Editor.
This changes the Compose preview so it does not set it as default and it
also blacklists the Compose mode as default.

Fixes: 144829328
Test: Adds regression test
Change-Id: Id655ff97a36ac48bd6df0c17f8d51d87da8b3c15