Update constraint set and layout model if needed.
When adding a view from the palette into an existing motion layout
the selection in the constraint set table in the motion editor did
not reflect the newly added View.
The problem is that the table model did not yet contain the new
View. With this change: rebuild the table if the model does not
match the ids that it should contain.

Fixes: 142289813
Test: N/A

Change-Id: I28bfab17e8fd681921053af06e7479709cd80500