Add the Theme selector menu for Drawable files
Sometimes, files can be opened in the 'Material.Light' theme. Likely due
to b/73332530.

For layout files, the user can simply select the correct theme in the
Layout Editor, but the same cannot be done for Drawables.

This adds the Theme selector menu for Drawable files, so that it's
possible not only to fix Drawable previews in the Layout Editor but also
preview Drawables under a different theme.

Bug: 145026814
Test: manual
Change-Id: I072ebd97196691706871afe4e93ab9e2c915be73