[Dsl,.kts] fix adding to single-element lists
When we parse a single-argument method call in KotlinScript, we end up
with a GradleDslLiteral associated with the (single) KtValueArgument,
which is reasonable.  However, when the property is actually a list
property, we might later wish to add to the list: a case which was not
handled by the KotlinScript writer.  Add a case to find the
KtValueArgument's parent as the holder for new elements to be added.

Fixes: 143762234
Test: four Dsl tests now enabled for KotlinScript
Change-Id: I5cbefdbcd055ff703e976f8fe05c7fc74434edeb