Don't use "compiled 9-patches" in the repository used by code editor
Native layoutlib relies on the fact that the framework resources used for
rendering (and distributed with Studio) contain compiled versions of 9-patch
drawables. FrameworkResourceRepository has support for swapping source files
of such resources, so that the right files are returned to layoutlib. The bug
was that this support was enabled also for the FrameworkResourceRepository
instance used by the code editor that is backed by SDK data, not
framework_res.jar used for rendering, which meant the swapped source files
didn't exist and were highlighted as missing.

Fixes: 144786008
Test: AndroidLayoutDomTest.kt
Change-Id: Ifc862a24fc10fdf08dc119f945bd5667f4e90362