Tweak AndroidMavenImportFixTest to reduce flakiness
Currently we're seeing some failures related to root changes being detected
during the call to doHighlighting which is illegal. Manually calling
doHighlighting is no longer needed in this test so I removed it.o
checkHighlighting is used in other Gradle test cases without issues, so
hopefully this will fix the issue.

Bug: None
Test: AndroidMavenImportFixTest.kt
Change-Id: I44c84b8daf01aa1d1d4811787e2d51fe0df5edea