[Dsl] map property autoexistence
Change the strategy for handling Dsl property elements which are maps:
instead of creating them in the block lookup or when they are
mentioned as properties, have them already present (but in a special
default state) in the Dsl elements themselves.

This is required to support direct setting of map entries with no
prior explicit reference to the map, for example a build.gradle.kts
file consisting only of:

android.buildTypes.create("foo").manifestPlaceholders["bar"] = "baz"

(and similar in Groovy, with explicit ['bar'] dereferencing.)

A number of tests have to be adjusted, to reflect the difference
between the previous implementation (map properties being absent) and
the new one (an empty map being present), but those test changes have
no effect on usage of the Dsl model/api.

This change also allows yet more simplification of
GradleDslFile.getBlockElement, because we no longer have to
automatically create maps there in order to attempt to handle direct
setting using dotted notation in Groovy.

Fixes: 143761530
Test: update existing tests
Change-Id: I943042c87e0f8e8073958b29bdef535f3fe2e17f