Do not access project in new module for new project
Because it doesn't exist

It is a temporary fix to make new project work, the better way will be
investigated later.
The core reason is that we create a new module module before creating a
project and then modify it with a project data. It should actually be
created only after creating a project but this is not easy to change.

Fixes: 145501733
Test: existing UI tests
Change-Id: Iabfb3bb96a02b1d469a281f6e88f5d0de81e4768