Improve shortcuts for changing split editor mode
Link the shortcuts for changing the split editor mode to the IntelliJ
actions for selecting next/previous tabs in multi-editor files. These
actions were used before when the design editors used to open resource
files using IntelliJ's default XML editor in addition to the design
editor itself. By linking the shortcuts to the aforementioned actions,
users can now change them using the settings panel.

Also, we're now displaying the shortcuts as the action tooltips. The
shortcuts change when the mode changes since the selected mode does not
have a shortcut assigned to it.

Bug: 144345936
Test: DesignToolsSplitEditorTest#testKeyboardShortcuts
Change-Id: Ifb7348e087345b820c7c526175122d64eb9fcafd