[Res Explorer] Add Refresh Previews button to toolbar
Additional to the right click menu action to refresh the preview of a
particular preview. This now adds a button on the toolbar to refresh all
resources being displayed in the ResourceExplorer, specifically, all
resources being shown of a particular ResourceType. Eg: Refresh all
Drawable previews.

The added button is disabled for resources which previews are updated
automatically, since those are usually more simple to preview. Eg:
colors, fonts, resource values.

Bug: 144722957
Test: ResourceExplorerListViewModelImplTest
Change-Id: I041ce852f832403e791d3353357ab5ea5fb0974b