[Dsl,.kts] ModelMapPropertyImpl tests
Implement the ModelMapPropertyImpl tests for KotlinScript.

A possibly-still-current bug about map keys having the same name as
property names led to the tests being de-symmetrised in two different
ways (renaming the keys or renaming the values).  Break the symmetry
in the same way in each test, so that we don't have to have four
almost-identical test files.

Bug: 138277630
Test: run the tests for KotlinScript
Change-Id: I4572e48a336f1e9a75d734cdca728d5a70a53c56