Get buildable of Java module from applied plugins
Studio used to check if a Java module is buildable by looking
for "compileJava" task. The task list in GradleProject will be
empty and can not be used for this purpose.
Instead, check directly if the Java plugin is applied to current
module, which is more reliable. The module is buildable only if
Java plugin is applied.

Bug: NA
Test: Covered by current comparison tests.
Change-Id: I2b14c502429040f73b982ee6ebe4875a4f17d95e