Release image buffers when deactivating NlModel
When selecting a different resource file, we deactivate the NlModels
belonging to it, but each model was still holding a reference to the
RenderResult image buffer, which in local tests took up to ~12MB of memory
for a reasonably complex layout. The NlModel was also keeping a reference to
RenderTask, which in turn keeps its own BufferedImage that could take an
extra ~8MB of memory, totalling ~20MB.

External OutOfMemory reports show users hitting ~400MB of memory usage
when 19 design editor tabs are open, and 1.27GB when 43 editors are open.

Now, we're disposing both RenderTask and RenderResult when changing files,
getting rid of the extra memory if the file is not showing. Consequently,
we're requesting a render when activating the model, which doesn't seem
to cause too much impact. If we detect that's causing a massive slowdown
to open resource files in the future, we can revisit the issue and
potentially implement a LRU mechanism to keep the last X opened buffers
in memory.

Fixes: 143305688
Fixes: 145803736
Bug: 143308628
Test: Manual. Profiling Studio shows the buffers are not kept in memory
      anymore, and the models load normally (without a massive slowdown)
      when changing files.
Change-Id: Ibf0089dc704a79f0d43145f2860776f22f12ce4b