[Res Explorer] Don't add 'ic_' prefix when importing
Don't add an 'ic_' prexif on the ResourceImportDialog, while it would be
nice to be consistent with the naming of VectorAssetAction, adding the
prefix here could be incorrect.

1) Not all images are necessarily considered icons by the user.

2) Modifying the name could lead to name collisions when importing
multiple images. Eg: importing "ic_my_image" and "my_image" both files
would end up with the same "ic_my_image" which would be incorrect.

Bug: 137210432
Test: updated
Change-Id: I7c46e0e1aa8cfb41e3802e81f898b097f3143113