[Dsl,.kts] support missingDimensionStrategy
The current implementation of missingDimensionStrategies as a parsed
Dsl element is of multiple individual missingDimensionStrategy
elements represented as ExpressionList objects.

This is the natural implementation for Groovy application statements,
where we represent source of the form

  missingDimensionStrategy 'foo', 'bar'

as a Dsl ExpressionList property with name "missingDimensionStrategy".
However, it fails with method call syntax, which is not too much of a
problem in Groovy but is the only way of expressing calls in
KotlinScript.

Add a clause in addParsedElement() to convert between the MethodCall
representation and the ExpressionList representation, which should fix
Groovy method call syntax and definitely does allow us to parse and
represent KotlinScript method calls.

(The existing representation also probably fails for
missingDimensionStrategy application statements with zero fallbacks,
as then the single argument would be interpreted as a Literal rather
than a MethodCall.)

Bug: 138277630
Test: enable missingDimensionStrategy tests for KotlinScript
Change-Id: I2205e76e2c86a7604666836e64bd96a8d450489f