Add termination listener to remote debugger
When the remote debugger terminates, it signals to the remote VM to
shut down the connection, which is the very same JDWP connection used
by DDMLib to communicate to the remote process. When this connection
goes down, both MonitorThread and Debugger tries to read from the
connection. Depending on which one goes first, it causes different
behavior:
MonitorThread - shuts down the Client leading to the bug.
Debugger - reopens the socket, things continue to be fine.

The fix is to attach a listener to the RemoteDebugProcessHandler to
get notified when the process handler terminates. When the process
handler terminates, we notify the Client to reopen its communication
channel to the process.

Fixes: 37104675
Test: manual as per bug, cannot repro after this CL
Change-Id: I7ce79c1403b98ea8a092fc5bb227df0ee3cb2cdf