[Dsl,.kts] special-case for dependencies
In the KotlinScript Dsl, there are a number of pre-defined
configurations, for which configuration extension functions are
provided by Gradle.  For all other configurations, the expected way of
specifying a dependency in that configuration is by invoking the
string name of that configuration with the description of the
dependency as an argument.

What this means is that when writing we have to check whether we are
writing a known configuration or not, and if not wrap the method name
in quotes; when parsing we must extract the configuration name from
quotes, if present.

Bug: 138277630
Test: existing (new tests forthcoming)
Change-Id: Ia4e81a790cdd2a685488c3da6ebab635f53eb733