Migrate DatabaseInspectorController to use coroutines
It shows how a "middle" layer can be converted from futures to
suspending functions and coroutnes:

- The controller consumes futures from a lower layer,
SqliteEvaluatorController, by suspending until they are resolved.

- It is also called from a higher layer,
DatabaseInspectorProjectServiceImpl, which means the project service
needs to pick the right scope for Kotlin's "structured concurrency".
Since the project service itself is not Disposable, we use a scope tied
to the Project (which is equivalent).

- The project service API is still based on ListenableFuture, showing
that the transition doesn't have to be done all at once. We use
kotlinx-coroutines-guava helper library to create futures tied to
coroutines, which is an implementation detail of
DatabaseInspectorProjectServiceImpl.

- Switching the controller to suspending functions doesn't require it to
manage scopes or implement AndroidCoroutineAware, this is done by the
caller (see above). But listeners registered by the controller schedule
asynchronous work in response to UI actions, which means they have to
pick a right scope. We use one tied to the Disposer lifecycle of the
controller, which seems most natural.

- Migrating to coroutines made it more explicit that listener methods
return immediately and only schedule tasks for later. Before the
migration these tasks were never cancelled.

- Suspending functions in the controller can enforce their own threading
requirements using `withContext`, so the caller no longer has to worry
about it.

- Since a lot of our tests run on the UI thread, we need to be careful
when invoking the suspending functions to avoid deadlocks. There's a new
utility in AndroidTestUtils for that.

Bug: None
Test: Existing tests should keep passing
Change-Id: Ifb68f0eb7b4f539d1b4762c8357846f523069a20