Extract lint IDE code into a separate module
In addition to code isolation which helps isolate dependencies
and makes running only affeced tests faster etc, this helps
make it possible to build a plugin for Lint that isn't Android
specific, such as for use with Kotlin or for Studio development.

This CL moves most of the generic lint support out to a separate
lint module (lint/ under tools/adt/idea, but called intellij.lint
in the UI), and leaves the Android specific checks and setup
in the android module. There's a new extension point used from
the lint support to look up necessary services, such as the issue
registry. It then renames the Android specific classes, such that
LintIdeClient is now the LintClient in the IDE, and
AndroidLintIdeClient is the specialization of the client when
running from the Android plugin. At the same time I took the
opportunity of renaming some other classes, such as "State", to
something more suitable.

Test: Existing
Bug: 145316442
Change-Id: I7351ab918b98114e20d8067a794766a995e89975