Update LaunchUtils.isWatchFaceApp to match current template
The function in question is used to determine whether the application is
a watch face application, in which case default run configuration is
configured to do nothing after installing an app, as opposed to
launching a default activity.

Currently, the function fails to recognize the current watch face
template, which results application freshly created from the template
fail to run with default settings.

This patch fixes two discrepancies between the template and matching
function:
1. Matching function expects project to not have any activities at all,
   but in the template instantiation there is a GoogleApiActivity coming
   from play-services-base library. Since it is not exported, the check
   has been changed to check whether there an exported activity, rather
   than no activities at all.
2. The check for a WallpaperService intent-filter expects service to
   have only one children, which is not the case for the current
   template. Instead, all the subtags are iterated over and checked for
   a particular action and category names.

Test: n/a
Bug: n/a
Change-Id: I7bd12439c1e4fab2ab4dc5906c4532b5855cbc11