Add default split editor mode comboboxes to settings panel
Add two comboboxes to the Layout Editor section of the settings panel: one
corresponding to drawable files, and another one for other resources. These
comboboxes can be used to set the default mode of the split editor when
opening resource files.

Files opened for the first time, and consequently new files, will be open in
the mode determined by the comboboxes. Files will continue be reopened in
the mode they were before being closed last time.

Bug: 147066397
Bug: 144513660
Test: Manual by changing the combobox value and opening new resource files.
Change-Id: I3413c4117ca2470f86156513f4efa3299fdabb09