[Res Manager] Fix 'Show In Resource Manager' misssing from project view
Seems that the ShowThumbnails action is being slowly removed from
IntelliJ, which caused the 'Show In Resource Manager' action to be
removed from the right click context menu in the project view.

This removes the dependency of the action to the ShowThumbnails action
and adds the Resource Manager action to the groups that it used to be
manually.

Fixes: 146773243
Test: manual
Change-Id: I8ea6b69af21e41af62fa169df3814e4e29a69d7e