147039508: Quickfix for replacing dynamic versions is broken
In ag/9225933 the sdkHandler parameter was marked non null
while there are several calls (including the this quickfix)
passing null to it which made the quickfix abort instead of
attempting to fix it.

There was a unit test which should have caught this but the
test gracefully degraded when SDK install doesn't have the
m2repository installed, which is no longer done by default.
Added a new test which uses the ability to instrument a
test-only offline repository instead.

Fixes: 147039508
Test: Unit test added

Change-Id: Ib9ca2bd2c901bb4ec7c919760080ae4315eee395