Separate NamedIdeaSourceProvider and IdeaSourceProvider.
This is to separate usages where the source provider name assigned by
the user matter from the cases where the set of source code
directories is what matters.  This allows further unification and
simplification of the later use.

The named source provider use case seems to be mostly limited Android
project view and merged manifest contributors.

Bug: n/a
Test: existing
Change-Id: Ic19bf3b85f01ad5a3a7957c002b8241c1c087e93