Fix two issues with Flamechart interaction in CpuAnalysis
The first issue is when selecting an event no data shows.
This was because of a bug where we selected the capture range instead
of the range of our selection.

The second issue was a bug when viewing the flame chart and moving the
selection did not update the flame chart. This is a subtle bug caused
by not holding onto a reference of the FlameChartView

Test: Manual
Fixes: 146354354
Change-Id: Ie9859867fcb843a6b7bdefe9824bc4aa7619fed4