Add sequential rendering function in DesignSurface
The sequential rendering is also used by default in DesignSurface.
Before the render requests are not started sequentially, but layoutlib
doesn't support parellel rendering so the renderings are still
executed sequentially. Thus the behaviour is same as before.

Although the behaviour is same as before, but it gives the chance to
interrupt the queued render requests in the following CLs.

This resummits a revert CL: Ia74a7e672ce66928b6e26c33197e076ddf00508f

Bug: 143308628
Bug: 148625501
Test: manually
Change-Id: If0465c1765a2e045e64bcbf4591b5e71146aab4d