Discover processes by listening for them
Prior to this change, we use information obtained from LaunchTask to
query transport for the process to discover. The biggest drawback of
this approach is the existence of race between AppInspection's discovery
mechanism and transport pipeline's own discovery mechanism.

With this change, we no longer rely on LaunchTask. We instead listen to
transport's STREAM and PROCESS events, which are fired when they start
and stop.

Introduce new class AppInspectionDiscoveryManager which listens for,
keeps track of, and establishes new connections.

Bug: 146664532
Test: AppInspectionDiscoveryManagerTest, manually tested sql inspector
Change-Id: Ia442c4b29e5eaed9b7e0160ca7834c37c6a6df51