Disable controls for new features on old devices
Previously we were hiding them, which could cause confusion. Instead we
now show them as disabled with an appropriate tooltip.
This change also includes some refactoring of
LayoutInspectorTransportRule to cover more cases and be simpler to use.
Also making TransportService.getChannelName() a constant, since as it
was we were needlessly creating the TransportService during some tests
and then not cleaning it up properly afterward.

Test: Added tests
Bug: 146437831
Change-Id: I3f0858c8a13cd7d63596533be560eaf90ce1fcb4