Don't leak AndroidFacet in MergedManifestSupplier
MergedManifestManager is a module service which maintains a reference
to a MergedManifestSupplier, which in turn references that module's
AndroidFacet. MergedManifestSupplier is a child of the facet in the
disposer tree, but MergedManifestManager isn't Disposable, and so this
causes test failures when the leak checker detects that an undisposed
MergedManifestManager is holding a reference to a disposed
MergedManifestSupplier, thereby leaking the facet.

This CL resolves the issue by making MergedManifestManager a
Disposable. Since it's a module service, it will automatically be
registered as a child of the module in the disposer tree. This CL
also updates MergedManifestManager's constructor so that it
registers MergedManifestSupplier as its child.

Since there seems to be some question of whether a facet's lifetime
matches that of the corresponding module, this CL also updates
MergedManifestSupplier to maintain a reference to the module and request
the AndroidFacet as needed. (Another alternative would have been to
make MergedManifestManager an AndroidFacetScopedService, but this
method is less complicated).

Fixes: 139940670
Test: MergedManifestManagerTest still passes + leak checker in PSQ
Change-Id: I88624acbbb5d7e2e21afa37056b63560932b3a85