[Res Explorer] Refactoring DrawableIconProvider
Refactoring DrawableIconProvider to -> SlowResourcePreviewManager, this
class now has no knowledge on the logic behind resource rendering, it
now only manages when to ask for a preview and what to do with it (like
scaling).

For rendering, SlowResourcePreviewManager instead takes a
SlowResourcePreviewProvider which provides a placeholder Image and a
CompletableFuture for the preview.

This allow us to make it easier to support previews for complex
resources, like generating a preview for Navigation files.

DrawableSlowPreviewProvider now has the logic to resolve Drawable
resources and generate previews. Similar for LayoutSlowPreviewProvider.

For Drawables we now show a simple icon as placeholder while the
Drawable is being rendered, it's the same icon used to represent the
Resource Manager.

For Layouts we show a placeholder icon that shows a card-like android
layout using very simple shapes.

Bug: 147157941
Test: updated
Change-Id: I91313f9092d0d688d99b22bb69687c54e092379d