Create ComposePreviewRunConfigurationProducer
Create a producer of ComposePreviewRunConfiguration for Compose Previews.
The producer essentially maps a PSI location (context) to a configuration,
which is initially named after the @Composable function and its fully
qualified name is also properly set.

The context where the ComposePreviewRunConfiguration is created from can
be any descendant of the @Composable function in the PSI tree, such as
its annotations, function name or even the keyword "fun".

Bug: 144500928
Test: ComposePreviewRunConfigurationProducerTest
Change-Id: Ib671af321d02046049e9f22c096913e6b9c5a28c