Connect to discovered processes lazily
Formerly, as new processes come online, AppInspectionDiscovery
immediately connects to them and broadcast their presence to
listeners. Connecting to a process incurs a non-trivial amount of
overhead on the device, so it's more prudent to connect lazily.

Now, AppInspectionDiscovery will track ProcessDescriptor only. And it
exposes an API to allow users to obtain an AppInspectionTarget if they
provide a descriptor.

AppInspectionJarCopier is the wrapper we use to wrap intellij platform
dependencies (TransportFileManager which depends on message bus, and
disposable), and pass to our API. It used to be created proactively by
the discovery mechanism, which is why part of our discovery code is in
the IDE module (AppInspectionDiscoveryManager). In this CL, the
responsibility of creating AppInspectionJarCopier and of requesting
AppInspectionTarget  is relegated to the user of discovery. Therefore,
AppInspectionDiscoveryManager is moved to API module and merged with
AppInspectionDiscoveryHost. So now, the "discovering" is rightfully done
in api.

Test: AppInspectionDiscoveryHostTest, AppInspectionDiscoveryTest
Bug: 147321791
Change-Id: I7cbb7afc4702a4369e3ace6153debba6ce2f1c8c