[sync] be more tolerant with gradle wrapper dist directories
The gradle wrapper constructs directories of the form

  .../wrapper/dists/gradle-<version>-bin/<hash>/gradle-<version>/...

where <version> is some version identifier, and <hash> is the base-36
representation of the 128-bit quantity representing the md5sum of the
URI of the distribution (see rootDirName() and getHash() in
PathAssembler.java in subprojects/wrapper in gradle source).

Most of the time, the base-36 representation of a 128-bit quantity is
25 characters long.  However, roughly 6% of the time, it will be less
than that, because leading 0s are not printed.  By relaxing the length
constraint on the regular expression for this directory to allow
strings between 22 and 25 characters (inclusive), we reduce failures
from this cause to around one time in one million.

Bug: n/a
Test: less likely to have a spurious failure
Change-Id: I4dcebf179b06ea5c56f749e9214b6dfa4585dccd