Fix flaky ResourceFolderRepositoryTest.testBitmapUpdated
Because UIUtil.dispatchAllInvocationEvents was invoked before the
condition would be checked, there was a race condition that could cause
that some events would still be in the queue when checking the
condition. In those cases, the test could fail.

The fix moves the dispatchAllInvocationEvents to the condition so the
events are always processed before the condition checked and it is
invoked multiple times.

Fixes: 135937224
Test: ResourceFolderRepositoryTest.testBitmapUpdated
Change-Id: I60b59a8e8320ccbbeb54d09b46e1f92234a51d8a