[Gradle Dsl Model] Correctly filter default properties.
With the newly introduced DEFAULT elementState, when getting properties
of a dsl element, we filter those that are default, but these might
have their state changed later (ex: if they are modified and need to be
added to the file).
Add isDefaultElement() to ElementItem that consider an element as
default when it has no children element.

Bug: N/A (This bug was introduced by a following CL)
Test: all tests passes
Change-Id: I1dde4b8c552d206d14db85aae2df8de58e361c2c