[Dsl] rework RepositoryModel.getType()
For URL-based repositories, it is the url's identity that matters,
rather than the particular Dsl object; a jcenter repository defined
with jcenter() and a repository defined with maven { ... } with a url
property pointing to https://jcenter.bintray.com/ should be
functionally equivalent.

Really I wanted to delete RepositoryModel.getType(), because (apart
from the distinction between flatDir and url-based), it doesn't really
make sense to consider these things as types, but unfortunately
RepositoryModel.getType() is part of the public Dsl API.  Instead,
rework the implementation to consider the url explicitly.

Bug: 144679123
Test: existing PsModule tests, others updated
Change-Id: I5083f5418f5a0d99562bebbe8b1517985208aa1e