Put focus on the editor when in Code or Split mode
When changing modes in the Split Editor, we were focusing on the parent
component, which contains both the preview and the text editor. It makes
sense, however, to put the focus on the Editor of the TextEditor whenever
we can edit some text, i.e. in "Code" or "Split" mode.

This CL makes that change, setting the focus to SplitEditor.myEditor.editor
when in Code or Split modes, and to myPreview when in Design mode.

Fixes: 146604992
Test: Manual. Change modes and notice the focus on the editor.
Change-Id: I60ca2330a6c2060e690d80f9e4453ee5713add57
(cherry picked from commit 9d186120788b4daf4b212a3804b57da6072ba7a3)