Remove obsolete early abort in databinding code
In a previous incarnaction of an algorithm in our codebase, it used to
rely on the resource repository manager for some information about
resources across the whole app, aborting if the information wasn't ready
yet.

However, we've since migrated much of our codebase over to using a custom
index, except this old check got left behind. This can cause us to
generate an incorrect light binding class that then gets cached.

A concrete example:

<include android:id="@+id/target"
         layout="@layout/target_layout"
         ... />

should create a field "target: TargetLayoutBinding"

but, if app resources aren't available at the time of parsing, even
though we have everything we need to generate that field, we end up
aborting early and creating the fallback "target: View" instead.

Fixes: 146538006
Test: Covered by existing
Change-Id: I1bf688721df0be46cae26298f01ff34e7853bc80