Only test library versions of 2 activities
Mostly isLibrary is used in module templates which is not properly
covered in template test at the moment (and probably not worth it).

There are only 5 cases of usage it in other templates and 3 of them
(sharedManifestMacros, androidTvActivity and BlankWearActivity) will
still be fully covered because they have
`if (isLaucher && !isLibrary) { ... }`
and we cover both values of isLaucher.

Other 2 are still covered by ad-hoc code.

It should significantly speed up TemplateTests with no decrease in code
coverage.

Bug: 67313506
Test: this is a test

Change-Id: I0b633520364432a34c99c6d7438c3dab0d53029a